Whitelist calls to pure functions from exception list
If a function is marked as pure function, we do not
except it to throw an exception. Most of the time
functions from this list are either inlined or NOP calls.
Return false to the query asking if these calls can raise
exception so that optimizations is not restricted on blocks
containing these calls.

Signed-off-by: Rahil Shah <rahil@ca.ibm.com>