Fix postGRA Block splitter
PassThrough under PassThrough node under GlRegDeps was not expected by
post GRA Block splitter. It is possible to have such node in trees which
is currently generated by RegDepCopy Removal optimization to minimize
register spillings when there multiple registers used for same node.
This commit fixes the block splitter so that when it encounters such node,
it traverses through the child to get actual node.

Signed-off-by: Rahil Shah <rahil@ca.ibm.com>