Clean-up code related to AlreadySignExtended register flag
In https://github.com/eclipse-omr/omr/pull/8035 we introduced a new
node flag to notify the code-gen that the node was sign extended when
evaluated. This removes the the use of AlreadySignExtended register
flag which can be propagated to other nodes if the register is reused
through clobber evaluation of the node.
This commit includes the clean-up related AlreadySignExtended register
flag.

Signed-off-by: Rahil Shah <rahil@ca.ibm.com>